[git-p4: depot-paths = "//open/dt/dev/": change = 14028]

+++ fennel/common/FileStatsTarget.cpp
@@ -1,10 +1,10 @@
 /*
 // $Id$
 // Fennel is a library of data storage and processing components.
-// Copyright (C) 2005-2005 The Eigenbase Project
-// Copyright (C) 2005-2005 Disruptive Tech
-// Copyright (C) 2005-2005 LucidEra, Inc.
-// Portions Copyright (C) 1999-2005 John V. Sichi
+// Copyright (C) 2005 The Eigenbase Project
+// Copyright (C) 2005 SQLstream, Inc.
+// Copyright (C) 2005 Dynamo BI Corporation
+// Portions Copyright (C) 1999 John V. Sichi
 //
 // This program is free software; you can redistribute it and/or modify it
 // under the terms of the GNU General Public License as published by the Free
@@ -40,12 +40,12 @@ std::string FileStatsTarget::getFilename() const
 void FileStatsTarget::beginSnapshot()
 {
     assert(!filename.empty());
-    snapshotStream.open(filename.c_str(),std::ios::trunc);
+    snapshotStream.open(filename.c_str(), std::ios::trunc);
 
     // TODO:  re-enable this.  I disabled it since /tmp/fennel.stats
     // can't be opened on mingw; need to parameterize it better
     // (or put in Performance Monitor integration)
-    
+
     // assert(snapshotStream.good());
 }
 
@@ -54,7 +54,7 @@ void FileStatsTarget::endSnapshot()
     snapshotStream.close();
 }
 
-void FileStatsTarget::writeCounter(std::string name,uint value)
+void FileStatsTarget::writeCounter(std::string name, int64_t value)
 {
     snapshotStream << name << ' ' << value << std::endl;
 }
